The 'dude you're using tab-based indents' commit
Well this is a fun one! Enjoy the results of IntelliJ taking out entire project and reformatting every file in it. Enjoy such pleasures as...
- Tab based indents replaced with 4 spaces, project-wide
- Adding indents where they're missing, or inadequately indented
- Adding line breaks for clearer code viewing

Look forward to round 2, known as 'The Cleanup'. Features such as:
- Burning all one line if statements at the stake
- Throwing Runtime exceptions instead of returning null when something should never happen
- Adding curly braces to all if statements (because Alastair said to be scared if not)
- Never before seen levels of passive aggressiveness

All this, coming soon to a repo near you!

TL:DR; if you don't pull this the merge request is going to be bad times.